Fine Art personal statement opinions?
I perceive art as a way of understanding life and understanding myself.
The most intriguing aspect of the foundation course at thus far is how I have developed my approach to the creation of art. To let the exploration of ideas become the focus of work and not to think of a project in terms of a path towards a final piece. This has encourages me to think more broadly about how I develop work.

https://www.tejar.pk/dyson-airblade-db-hand-dryer